turns 90 degrees to the right (from facing 0 degrees to 270 degrees), moves toward red alliance wall
