--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/opmode/austinPath.java
@@ -1,34 +1,0 @@
-package org.firstinspires.ftc.teamcode.drive.opmode;
-
-import com.acmerobotics.roadrunner.geometry.Pose2d;
-import com.acmerobotics.roadrunner.geometry.Vector2d;
-import com.acmerobotics.roadrunner.trajectory.Trajectory;
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-
-import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
-
-import java.util.Vector;
-
-@Autonomous (group = "drive")
-
-public class austinPath extends LinearOpMode {
-
-    @Override
-
-    public void runOpMode() {
-
-        SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
-
-        Trajectory driveUnderObstacle = drive.trajectoryBuilder(new Pose2d(-9, 59, Math.toRadians(0)))
-                .lineTo(new Vector2d(30, 59))
-                .build();
-
-        waitForStart();
-
-        if(isStopRequested()) return;
-
-        //drive.turn(Math.toRadians(-90));
-        drive.followTrajectory(driveUnderObstacle);
-    }
-}
